+++ packages/api-i18n-ddb/src/operations/locales/fields.ts
@@ -1,0 +1,15 @@
+import { LocaleDynamoDbFieldPlugin } from "~/plugins/LocaleDynamoDbFieldPlugin";
+
+export default () => [
+    new LocaleDynamoDbFieldPlugin({
+        field: "code"
+    }),
+    new LocaleDynamoDbFieldPlugin({
+        field: "createdOn",
+        type: "date"
+    }),
+    new LocaleDynamoDbFieldPlugin({
+        field: "createdBy",
+        path: "createdBy.id"
+    })
+];