+++ packages/api-page-builder-so-ddb-es/src/operations/pageElement/fields.ts
@@ -1,0 +1,13 @@
+import { PageElementDynamoDbElasticFieldPlugin } from "~/plugins/definitions/PageElementDynamoDbElasticFieldPlugin";
+
+export default () => [
+    new PageElementDynamoDbElasticFieldPlugin({
+        field: "createdOn",
+        type: "date"
+    }),
+    new PageElementDynamoDbElasticFieldPlugin({
+        field: "createdBy",
+        path: "createdBy.id",
+        sortable: false
+    })
+];