+++ packages/api-security-admin-users-so-ddb/src/plugins/UserDynamoDbFieldPlugin.ts
@@ -1,0 +1,5 @@
+import { FieldPlugin } from "@webiny/db-dynamodb/plugins/definitions/FieldPlugin";
+
+export class UserDynamoDbFieldPlugin extends FieldPlugin {
+    public static readonly type: string = "adminUsersSecurity.dynamodb.field.user";
+}